+++ src/kasperl/filter/__init__.py
@@ -4,6 +4,7 @@ from ._attach_metadata import AttachMetaData
 from ._block import Block
 from ._check_duplicate_filenames import CheckDuplicateFilenames, DUPLICATE_ACTIONS, DUPLICATE_ACTION_IGNORE, DUPLICATE_ACTION_WARN, DUPLICATE_ACTION_DROP, DUPLICATE_ACTION_ERROR
 from ._copy_files import CopyFiles
+from ._count_data import CountData
 from ._discard_by_name import DiscardByName
 from ._list_to_sequence import ListToSequence
 from ._log_data import LogData